(File) rename monthlyChart.tsx to periodicChart.tsx
