+++ src/features/Competition/utils/periodicChart.tsx
